+++ src/components/creator/ActorSelection/InProgressDialog.tsx
@@ -1,0 +1,43 @@
+import { Button, Dialog, DialogContent, Stack } from "@mui/material"
+import { useState } from "react"
+import { Link } from "react-router-dom"
+import { useTranslation } from "react-i18next"
+import theme from "../../../theme"
+import { LocalStorage } from "../../../localStorage"
+
+export const ChallengeInProgressDialog = () => {
+	const thereIsChallengeInCreation: boolean = !!LocalStorage.getCreatorChallenge()
+	const [openModal, setOpenModal] = useState(thereIsChallengeInCreation)
+	const { t } = useTranslation("creator")
+
+	const onDiscard = () => {
+		setOpenModal(false)
+		LocalStorage.saveCreatorChallenge(null)
+	}
+
+	return (
+		<>
+			<Dialog
+				scroll="paper"
+				maxWidth={false}
+				open={openModal}
+				onClose={() => setOpenModal(false)}
+				data-testid="challenge-progress-warning"
+			>
+				<DialogContent>
+					<Stack direction="column">
+						{t("selection.challengeBeingCreated")}
+						<b>{t("selection.discardWarning")}</b>
+						<Stack sx={{ marginTop: theme.spacing(1) }} direction='row' gap={theme.spacing(1)} justifyContent='flex-end'>
+							<Button variant="contained" color='error' onClick={onDiscard}>{t("selection.discard")}</Button>
+							<Link to="/creador/editar">
+								<Button variant="contained" color="success" >{t("selection.continueEditing")}</Button>
+							</Link>
+						</Stack>
+					</Stack>
+				</DialogContent>
+			</Dialog>
+		</>
+	)
+}
+
