--- src/components/creator/selection.module.css
@@ -1,68 +1,0 @@
-.selection{
-    text-align: center;
-    align-items: center;
-    justify-content: space-evenly;
-    display: flex;
-    background-color: var(--home-background);
-}
-
-/*
-.selection-text {
-    color: var(--home-text-color) !important;
-}
-*/
-.selection-textAnim {
-    animation: 1s textAnim;
-  }
-  @keyframes textAnim {
-    0% {
-      opacity: 0;
-      
-    }
-    25% {
-      opacity: 0.25;
-    }
-    50% {
-      opacity: 0.5;
-      
-    }
-    75% {
-        opacity: 0.75;
-        
-      }
-    100% {
-      opacity: 1;
-      
-    }
-  }
-
-/*
-.selection-textAnim {
-    animation: 2s textAnim ease-out;
-  }
-  @keyframes textAnim {
-    0% {
-      opacity: 0;
-      
-    }
-    20% {
-      opacity: 0.2;
-    }
-    40% {
-      opacity: 0.4;
-      
-    }
-    60% {
-        opacity: 0.6;
-        
-      }
-    80% {
-        opacity: 0.8;
-        
-      }
-    100% {
-      opacity: 1;
-      
-    }
-  }
-*/
